Move GcsBucket class to its own file.
This is part of the work to break the testbench script into smaller
pieces.
